+++ frontend/src/evac/EvacViews.js
@@ -1,21 +1,71 @@
 import React, { useEffect, useState } from 'react';
 import axios from "axios";
-import { Link } from 'raviger';
-import { Button, Col, Container, Form, FormCheck, OverlayTrigger, Row, Tooltip } from 'react-bootstrap';
+import { Link, navigate } from 'raviger';
+import { Form, Formik } from 'formik';
+import { Button, Col, Container, FormGroup, FormCheck, OverlayTrigger, Row, Tooltip } from 'react-bootstrap';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
   faBandAid, faCar, faClipboardList, faShieldAlt, faTrailer
 } from '@fortawesome/free-solid-svg-icons';
 import { CircleMarker, Map, TileLayer, Tooltip as MapTooltip } from "react-leaflet";
 import L from "leaflet"
+import { Typeahead } from 'react-bootstrap-typeahead';
 
+import 'react-bootstrap-typeahead/css/Typeahead.css';
 import 'leaflet/dist/leaflet.css';
+
 
 export function Dispatch() {
 
   const [data, setData] = useState({service_requests: [], isFetching: false, bounds:L.latLngBounds([[0,0]])});
   const [mapState, setMapState] = useState({});
   const [statusOptions, setStatusOptions] = useState({aco_required:false});
+  function getSelectedServiceRequests(){
+    var selectedIds = [];
+    for (var sr in mapState){
+      if (mapState[sr].hasOwnProperty('checked') & mapState[sr]['checked'] == true){
+      selectedIds.push(sr);
+      };
+    };
+    return selectedIds;
+  }
+
+  // Team Selector Code
+  const [selected, setSelected] = useState([]);
+  const [teamData, setTeamData] = useState({options: [], isFetching: false});
+  function getSelectedTeamMembers(){
+    var selectedIds = [];
+    selected.forEach(function(item){
+      selectedIds.push(item.id);
+    })
+    return selectedIds;
+  }
+
+  useEffect(() => {
+    let source = axios.CancelToken.source();
+    const fetchTeamMembers = async () => {
+      setTeamData({options: [], isFetching: true});
+      await axios.get('/evac/api/evacteammember/', {
+        cancelToken: source.token,
+      })
+      .then(response => {
+        var options = []
+        response.data.forEach(function(teammember){
+          options.push({id: teammember.id, label: teammember.display_name})
+        });
+        setTeamData({options: options, isFetching: false});
+      })
+      .catch(error => {
+        console.log(error.response);
+        setTeamData({options: [], isFetching: false});
+      });
+    };
+    fetchTeamMembers();
+    return () => {
+      source.cancel();
+    };
+  }, [])
+  // End Team Selector Code
 
   // Handle aco_required toggle.
   const handleACO = async event => {
@@ -126,115 +176,148 @@ export function Dispatch() {
   }, [statusOptions]);
 
   return (
-    <Container>
-      <Row>
-        <Col xs={12}>
-          <Map className="mx-auto d-block" bounds={data.bounds} onMoveEnd={onMove}>
-            <TileLayer
-              url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
-              attribution='&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors'
+    <Formik
+      initialValues={{
+        team_members: '',
+      }}
+      onSubmit={(values, { setSubmitting }) => {
+        values.service_requests = getSelectedServiceRequests();
+        values.team_members = getSelectedTeamMembers();
+        setTimeout(() => {
+          axios.post('/evac/api/evacassignment/', values)
+          .then(function() {
+            navigate('/evac');
+          })
+          .catch(error => {
+            console.log(error.response);
+          });
+          setSubmitting(false);
+        }, 500);
+      }}
+    >
+    <Form>
+      <Container>
+        <Row>
+          <FormGroup style={{ marginTop: '20px' }}>
+            <Typeahead
+              id="team-members"
+              multiple
+              onChange={setSelected}
+              options={teamData.options}
+              placeholder="Choose team members..."
+              selected={selected}
             />
-            {data.service_requests.map(service_request => (
-              <CircleMarker
-                key={service_request.id}
-                center={{lat:service_request.latitude, lng: service_request.longitude}}
-                color={mapState[service_request.id] ? mapState[service_request.id].color : ""}
-                fill={true}
-                fillOpacity="1"
-                onClick={() => handleMapState(service_request.id)}
-                radius={5}
-              >
-                <MapTooltip autoPan={false}>
-                  <span>
-                    {mapState[service_request.id] ? <span>{Object.keys(mapState[service_request.id].matches).map((key,i) => (
-                      <span key={key} style={{textTransform:"capitalize"}}>
-                        {i > 0 && ", "}{prettyText(key.split(',')[1], key.split(',')[0], mapState[service_request.id].matches[key])}
+          </FormGroup>
+          </Row>
+          <Row>
+            <Col xs={12}>
+              <Map className="mx-auto d-block" bounds={data.bounds} onMoveEnd={onMove}>
+                <TileLayer
+                  url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
+                  attribution='&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors'
+                />
+                {data.service_requests.map(service_request => (
+                  <CircleMarker
+                    key={service_request.id}
+                    center={{lat:service_request.latitude, lng: service_request.longitude}}
+                    color={mapState[service_request.id] ? mapState[service_request.id].color : ""}
+                    fill={true}
+                    fillOpacity="1"
+                    onClick={() => handleMapState(service_request.id)}
+                    radius={5}
+                  >
+                    <MapTooltip autoPan={false}>
+                      <span>
+                        {mapState[service_request.id] ? <span>{Object.keys(mapState[service_request.id].matches).map((key,i) => (
+                          <span key={key} style={{textTransform:"capitalize"}}>
+                            {i > 0 && ", "}{prettyText(key.split(',')[1], key.split(',')[0], mapState[service_request.id].matches[key])}
+                          </span>
+                        ))}</span>:""}
+                        <br />
+                        {service_request.full_address}
                       </span>
-                    ))}</span>:""}
-                    <br />
-                    {service_request.full_address}
+                    </MapTooltip>
+                  </CircleMarker>
+                ))}
+              </Map>
+            </Col>
+          </Row>
+          <Row>
+            <FormCheck id="aco_required" name="aco_required" type="switch" label="ACO Required" checked={statusOptions.ACORequired} onChange={handleACO} />
+            <Button type="submit" className="mt-2 mb-1">Deploy!</Button>
+          {data.service_requests.map(service_request => (
+            <div key={service_request.id} className="mt-2" hidden={mapState[service_request.id] ? mapState[service_request.id].hidden : false}>
+              <div className="card-header">
+                <span style={{display:"inline"}} className="custom-control-lg custom-control custom-checkbox">
+                  <input className="custom-control-input" type="checkbox" name={service_request.id} id={service_request.id} onChange={() => handleMapState(service_request.id)} checked={mapState[service_request.id] ? mapState[service_request.id].checked : false} />
+                  <label className="custom-control-label" htmlFor={service_request.id}></label>
+                </span>
+                {mapState[service_request.id] ?
+                <span>{Object.keys(mapState[service_request.id].matches).map((key,i) => (
+                  <span key={key} style={{textTransform:"capitalize"}}>
+                    {i > 0 && ", "}{prettyText(key.split(',')[1], key.split(',')[0], mapState[service_request.id].matches[key])}
                   </span>
-                </MapTooltip>
-              </CircleMarker>
-            ))}
-          </Map>
-        </Col>
-
-      </Row>
-      <Form>
-          <FormCheck id="aco_required" name="aco_required" type="switch" label="ACO Required" checked={statusOptions.ACORequired} onChange={handleACO} />
-          <Button type="submit" className="mt-2 mb-1">Deploy!</Button>
-        </Form>
-      {data.service_requests.map(service_request => (
-        <div key={service_request.id} className="mt-2" hidden={mapState[service_request.id] ? mapState[service_request.id].hidden : false}>
-          <div className="card-header">
-            <span style={{display:"inline"}} className="custom-control-lg custom-control custom-checkbox">
-              <input className="custom-control-input" type="checkbox" name={service_request.id} id={service_request.id} onChange={() => handleMapState(service_request.id)} checked={mapState[service_request.id] ? mapState[service_request.id].checked : false} />
-              <label className="custom-control-label" htmlFor={service_request.id}></label>
-            </span>
-            {mapState[service_request.id] ?
-            <span>{Object.keys(mapState[service_request.id].matches).map((key,i) => (
-              <span key={key} style={{textTransform:"capitalize"}}>
-                {i > 0 && ", "}{prettyText(key.split(',')[1], key.split(',')[0], mapState[service_request.id].matches[key])}
-              </span>
-            ))}</span>:""}
-            {service_request.aco_required ?
-            <OverlayTrigger
-              key={"aco"}
-              placement="top"
-              overlay={
-                <Tooltip id={`tooltip-aco`}>
-                  ACO required
-                </Tooltip>
-              }
-            >
-              <FontAwesomeIcon icon={faShieldAlt} inverse className="ml-1"/>
-            </OverlayTrigger>
-             : ""}
-            {service_request.injured ?
-            <OverlayTrigger
-              key={"injured"}
-              placement="top"
-              overlay={
-                <Tooltip id={`tooltip-injured`}>
-                  Injured animal
-                </Tooltip>
-              }
-            >
-              <FontAwesomeIcon icon={faBandAid} inverse className="ml-1"/>
-            </OverlayTrigger>
-             : ""}
-            {service_request.accessible ?
-            <OverlayTrigger
-              key={"accessible"}
-              placement="top"
-              overlay={
-                <Tooltip id={`tooltip-accessible`}>
-                  Easily accessible
-                </Tooltip>
-              }
-            >
-              <FontAwesomeIcon icon={faCar} inverse className="ml-1"/>
-            </OverlayTrigger>
-             : ""}
-            {service_request.turn_around ?
-            <OverlayTrigger
-              key={"turnaround"}
-              placement="top"
-              overlay={
-                <Tooltip id={`tooltip-turnaround`}>
-                  Room to turn around
-                </Tooltip>
-              }
-            >
-              <FontAwesomeIcon icon={faTrailer} inverse className="ml-1"/>
-            </OverlayTrigger>
-             : ""}
-            <span className="ml-2">| &nbsp;{service_request.full_address}</span>
-            <Link href={"/hotline/servicerequest/" + service_request.id}> <FontAwesomeIcon icon={faClipboardList} inverse /></Link>
-          </div>
-        </div>
-      ))}
-    </Container>
+                ))}</span>:""}
+                {service_request.aco_required ?
+                <OverlayTrigger
+                  key={"aco"}
+                  placement="top"
+                  overlay={
+                    <Tooltip id={`tooltip-aco`}>
+                      ACO required
+                    </Tooltip>
+                  }
+                >
+                  <FontAwesomeIcon icon={faShieldAlt} inverse className="ml-1"/>
+                </OverlayTrigger>
+                 : ""}
+                {service_request.injured ?
+                <OverlayTrigger
+                  key={"injured"}
+                  placement="top"
+                  overlay={
+                    <Tooltip id={`tooltip-injured`}>
+                      Injured animal
+                    </Tooltip>
+                  }
+                >
+                  <FontAwesomeIcon icon={faBandAid} inverse className="ml-1"/>
+                </OverlayTrigger>
+                 : ""}
+                {service_request.accessible ?
+                <OverlayTrigger
+                  key={"accessible"}
+                  placement="top"
+                  overlay={
+                    <Tooltip id={`tooltip-accessible`}>
+                      Easily accessible
+                    </Tooltip>
+                  }
+                >
+                  <FontAwesomeIcon icon={faCar} inverse className="ml-1"/>
+                </OverlayTrigger>
+                 : ""}
+                {service_request.turn_around ?
+                <OverlayTrigger
+                  key={"turnaround"}
+                  placement="top"
+                  overlay={
+                    <Tooltip id={`tooltip-turnaround`}>
+                      Room to turn around
+                    </Tooltip>
+                  }
+                >
+                  <FontAwesomeIcon icon={faTrailer} inverse className="ml-1"/>
+                </OverlayTrigger>
+                 : ""}
+                <span className="ml-2">| &nbsp;{service_request.full_address}</span>
+                <Link href={"/hotline/servicerequest/" + service_request.id}> <FontAwesomeIcon icon={faClipboardList} inverse /></Link>
+              </div>
+            </div>
+          ))}
+          </Row>
+      </Container>
+    </Form>
+  </Formik>
   )
 }
